+++ lib/features/home/domain/repository/home_base_repository.dart
@@ -93,10 +93,10 @@ abstract class HomeBaseRepository {
   Future<Either<Failure,void>> deleteExercisePlanDetails(DeleteExercisePlanDetailsParams params);
 
   Future<Either<Failure, AddNutritionEntity>> addNutrition({
-        required double fat,
-        required double carb,
-        required double protein,
-        required double calories,
+        required int fat,
+        required int carb,
+        required int protein,
+        required int calories,
         required String? howToPrepare,
         required Map component,
         required File nutritionPic,
